Title: 특정 문자로 시작하는 문자열, Time: 5ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 231214/특정 문자로 시작하는 문자열/strings-that-start-with-a-specific-character.c
@@ -19,6 +19,6 @@ int main() {
         if (c[i][0] == c1) co++;
     }
 
-    printf("%.0f %.0f.00", (float)co, (float)h / n);
+    printf("%d %.0f.00", co, (float)h / n);
     return 0;
 }
